refactor: RestTestApi class 패키지 위치 이동
