+++ src/test/java/nextstep/subway/utils/RestTestApi.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.acceptance;
+package nextstep.subway.utils;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
